Добавил модульность в проект
