+++ js/main.js
@@ -1,61 +1,6 @@
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
+/* eslint-disable no-console */
+import { getArrayPhotos } from './create-array-photo.js';
 
-const AUTHOR_NAMES = ['Павел','Владимир','Валерия','Марина','Максим','Кирилл'];
+console.log(getArrayPhotos());
 
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-
-const generateUserComment = () => {
-  let id = 1;
-
-  return () => {
-    const userComment = {
-      id,
-      avatar: `img/avatar-{${getRandomInteger(1, 6)}}.svg`,
-      message: getRandomArrayElement(COMMENT_MESSAGES),
-      name: getRandomArrayElement(AUTHOR_NAMES),
-    };
-
-    id++;
-
-    return userComment;
-  };
-};
-
-
-const createUserPhoto = () => {
-  let id = 1;
-
-  return () => {
-    const userPhoto = {
-      id,
-      url: `photos/{${id}}.jpg`,
-      description: `Очень красивая фотография под номером ${id}. Мне очень понравилось.`,
-      likes: getRandomInteger(15, 200),
-      comments: Array.from({length: getRandomInteger(1, 30)}, generateUserComment()),
-    };
-
-    id++;
-
-    return userPhoto;
-  };
-};
-
-
-// eslint-disable-next-line no-unused-vars
-const photos = Array.from({length: 25}, createUserPhoto());
